+++ Common/src/main/java/fuzs/strawstatues/client/gui/screens/strawstatue/StrawStatueModelPartsScreen.java
@@ -1,12 +1,14 @@
 package fuzs.strawstatues.client.gui.screens.strawstatue;
 
+import com.mojang.authlib.GameProfile;
+import fuzs.strawstatues.StrawStatues;
 import fuzs.strawstatues.api.client.gui.components.TickBoxButton;
 import fuzs.strawstatues.api.client.gui.screens.armorstand.ArmorStandTickBoxScreen;
 import fuzs.strawstatues.api.network.client.data.DataSyncHandler;
 import fuzs.strawstatues.api.world.inventory.ArmorStandHolder;
 import fuzs.strawstatues.api.world.inventory.data.ArmorStandScreenType;
-import fuzs.strawstatues.StrawStatues;
 import fuzs.strawstatues.network.client.C2SStrawStatueModelPartMessage;
+import fuzs.strawstatues.network.client.C2SStrawStatueOwnerMessage;
 import fuzs.strawstatues.world.entity.decoration.StrawStatue;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.Button;
@@ -34,6 +36,21 @@ public class StrawStatueModelPartsScreen extends ArmorStandTickBoxScreen<PlayerM
             strawStatue.setModelPart(option, value);
             StrawStatues.NETWORK.sendToServer(new C2SStrawStatueModelPartMessage(option, value));
         }, Button.NO_TOOLTIP);
+    }
+
+    @Override
+    protected void syncNameChange(String input) {
+        StrawStatues.NETWORK.sendToServer(new C2SStrawStatueOwnerMessage(input));
+    }
+
+    @Override
+    protected int getNameMaxLength() {
+        return 16;
+    }
+
+    @Override
+    protected String getNameDefaultValue() {
+        return ((StrawStatue) this.getHolder().getArmorStand()).getOwner().map(GameProfile::getName).orElse("");
     }
 
     @Override